Implement HashFunction

+++ src/lib/types.hpp
@@ -36,7 +36,7 @@
 STRONG_TYPEDEF(uint32_t, ChunkID);
 STRONG_TYPEDEF(uint16_t, ColumnID);
 STRONG_TYPEDEF(uint16_t, PartitionID);
-STRONG_TYPEDEF(uint32_t, HashValue);
+STRONG_TYPEDEF(uint64_t, HashValue);
 STRONG_TYPEDEF(uint32_t, ValueID);  // Cannot be larger than ChunkOffset
 STRONG_TYPEDEF(uint32_t, NodeID);
 STRONG_TYPEDEF(int32_t, CpuID);
